item update

+++ src/shop-item/shop-item.service.ts
@@ -1,12 +1,16 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { AppService } from 'src/app.service';
+import { CoreOutPut } from 'src/common/dtos/output.dto';
 import { ShopUser } from 'src/shop-user/entities/shop-user.entity';
 import { Repository } from 'typeorm';
 import {
   AddShopItemInputDto,
   AddShopItemsOutPutDto,
 } from './dtos/addShopItem.dto';
+import { GetItemInfoOutPutDto } from './dtos/get-iteminfo.dto';
+import { SellitemsOutPutDto } from './dtos/sell-items.dto';
+import { UpdateItemInputDto, UpdateItemOutPut } from './dtos/update-item.dto';
 import { ShopIreceipt } from './eitities/shop-ireceipt.entity';
 import { ShopItem } from './eitities/shop-item.entity';
 
@@ -20,7 +24,7 @@ export class ShopItemService {
     private readonly appService: AppService,
   ) {}
 
-  async getSellItems(user: ShopUser) {
+  async getSellItems(user: ShopUser): Promise<SellitemsOutPutDto> {
     if (!user?.sellerInfo) {
       return {
         ok: false,
@@ -73,8 +77,65 @@ export class ShopItemService {
     };
   }
 
-  async deleteItemById(id: number) {
-    const shopItem = await this.itemRepository.findOne({ id });
+  async updateItem(
+    user: ShopUser,
+    id: number,
+    input: UpdateItemInputDto,
+  ): Promise<UpdateItemOutPut> {
+    try {
+      const shopItem = await this.itemRepository.findOne(
+        { id },
+        { relations: ['sellUserInfo'] },
+      );
+
+      if (shopItem.sellUserInfo.id !== user.sellerInfo.id) {
+        return {
+          ok: false,
+          err: '소유자가 아닙니다.',
+        };
+      }
+
+      if (input.detailHtml) {
+        await this.appService.deleteShopItemImageByHtml(shopItem.detailHtml);
+      }
+
+      // update
+      const ok = await this.itemRepository.update(id, {
+        updateAt: new Date(),
+        ...input,
+      });
+
+      if (!ok) {
+        return {
+          ok: false,
+          err: '유효하지않은 접근 입니다.',
+        };
+      }
+
+      return {
+        ok: true,
+      };
+    } catch (err) {
+      return {
+        ok: false,
+        err: '유효하지않은 접근 입니다.',
+      };
+    }
+  }
+
+  async deleteItemById(user: ShopUser, id: number): Promise<CoreOutPut> {
+    const shopItem = await this.itemRepository.findOne(
+      { id },
+      { relations: ['sellUserInfo'] },
+    );
+
+    if (shopItem.sellUserInfo !== user.sellerInfo) {
+      return {
+        ok: false,
+        err: '소유자가 아닙니다.',
+      };
+    }
+
     await this.appService.deleteShopItemImageByHtml(shopItem.detailHtml);
 
     const result = await this.itemRepository.remove(shopItem);
@@ -90,7 +151,7 @@ export class ShopItemService {
     };
   }
 
-  async getItemById(id: number) {
+  async getItemById(id: number): Promise<GetItemInfoOutPutDto> {
     const item = await this.itemRepository.findOne(
       { id },
       { relations: ['sellUserInfo'] },
